+++ rmgpy/data/statmechfit.py
@@ -68,6 +68,7 @@ max_iter = 200
 
 ################################################################################
 
+
 def fit_statmech_to_heat_capacity(Tlist, Cvlist, n_vib, n_rot, molecule=None):
     """
     For a given set of dimensionless heat capacity data `Cvlist` corresponding
@@ -84,10 +85,12 @@ def fit_statmech_to_heat_capacity(Tlist, Cvlist, n_vib, n_rot, molecule=None):
     # You must specify at least 7 heat capacity points to use in the fitting;
     # you can specify as many as you like above that minimum
     if len(Tlist) < 7:
-        raise StatmechFitError('You must specify at least 7 heat capacity points to fit_statmech_to_heat_capacity().')
+        raise StatmechFitError("You must specify at least 7 heat capacity points to fit_statmech_to_heat_capacity().")
     if len(Tlist) != len(Cvlist):
-        raise StatmechFitError('The number of heat capacity points ({0:d}) does not match the number of temperatures '
-                               'provided ({1:d}).'.format(len(Cvlist), len(Tlist)))
+        raise StatmechFitError(
+            "The number of heat capacity points ({0:d}) does not match the number of temperatures "
+            "provided ({1:d}).".format(len(Cvlist), len(Tlist))
+        )
 
     # The number of optimization variables available is constrained to be less
     # than the number of heat capacity points
@@ -122,10 +125,15 @@ def fit_statmech_to_heat_capacity(Tlist, Cvlist, n_vib, n_rot, molecule=None):
     for i in range(n_rot):
         freq = hind[i][0]
         barr = hind[i][1]
-        inertia = (barr * constants.c * 100.0 * constants.h) / (8 * math.pi * math.pi * (freq*constants.c*100.0) ** 2)
+        inertia = (barr * constants.c * 100.0 * constants.h) / (8 * math.pi * math.pi * (freq * constants.c * 100.0) ** 2)
         barrier = barr * constants.c * 100.0 * constants.h * constants.Na
-        hr = HinderedRotor(inertia=(inertia * constants.Na * 1e23, "amu*angstrom^2"),
-                           barrier=(barrier / 1000., "kJ/mol"), symmetry=1, semiclassical=False, quantum=False)
+        hr = HinderedRotor(
+            inertia=(inertia * constants.Na * 1e23, "amu*angstrom^2"),
+            barrier=(barrier / 1000.0, "kJ/mol"),
+            symmetry=1,
+            semiclassical=False,
+            quantum=False,
+        )
         modes.append(hr)
 
     # Return the fitted modes
@@ -133,6 +141,7 @@ def fit_statmech_to_heat_capacity(Tlist, Cvlist, n_vib, n_rot, molecule=None):
 
 
 ################################################################################
+
 
 def fit_statmech_direct(Tlist, Cvlist, n_vib, n_rot, molecule=None):
     """
@@ -156,7 +165,7 @@ def fit_statmech_direct(Tlist, Cvlist, n_vib, n_rot, molecule=None):
     # Construct the initial guess
     # Initial guesses within each mode type must be distinct or else the
     # optimization will fail
-    x0 = np.zeros(n_vib + 2 * n_rot, np.float64)
+    x0 = np.zeros(n_vib + 2 * n_rot, float)
     # Initial guess for harmonic oscillator frequencies
     if n_vib > 0:
         x0[0] = 200.0
@@ -176,14 +185,13 @@ def fit_statmech_direct(Tlist, Cvlist, n_vib, n_rot, molecule=None):
 
     # Check that the results of the optimization are valid
     if not np.isfinite(x).all():
-        raise StatmechFitError('Returned solution vector is nonsensical: x = {0}.'.format(x))
+        raise StatmechFitError("Returned solution vector is nonsensical: x = {0}.".format(x))
     if igo == 8:
-        logging.warning('Maximum number of iterations reached when fitting spectral data for '
-                        '{0}.'.format(molecule.to_smiles()))
+        logging.warning("Maximum number of iterations reached when fitting spectral data for " "{0}.".format(molecule.to_smiles()))
     elif igo > 8:
-        logging.warning('A solver error occured when fitting spectral data for {0}.'.format(molecule.to_smiles()))
-    logging.debug('Fitting remaining heat capacity to {0} vibrations and {1} rotations'.format(n_vib, n_rot))
-    logging.debug('The residuals for heat capacity values is {}'.format(fit.evaluate(x)[0]))
+        logging.warning("A solver error occured when fitting spectral data for {0}.".format(molecule.to_smiles()))
+    logging.debug("Fitting remaining heat capacity to {0} vibrations and {1} rotations".format(n_vib, n_rot))
+    logging.debug("The residuals for heat capacity values is {}".format(fit.evaluate(x)[0]))
 
     # Postprocess optimization results
     vib = list(x[0:n_vib])
@@ -195,6 +203,7 @@ def fit_statmech_direct(Tlist, Cvlist, n_vib, n_rot, molecule=None):
 
 
 ################################################################################
+
 
 def fit_statmech_pseudo_rotors(Tlist, Cvlist, n_vib, n_rot, molecule=None):
     """
@@ -217,7 +226,7 @@ def fit_statmech_pseudo_rotors(Tlist, Cvlist, n_vib, n_rot, molecule=None):
     # Construct the initial guess
     # Initial guesses within each mode type must be distinct or else the
     # optimization will fail
-    x0 = np.zeros(n_vib + 2, np.float64)
+    x0 = np.zeros(n_vib + 2, float)
     # Initial guess for harmonic oscillator frequencies
     if n_vib > 0:
         x0[0] = 200.0
@@ -233,14 +242,13 @@ def fit_statmech_pseudo_rotors(Tlist, Cvlist, n_vib, n_rot, molecule=None):
 
     # Check that the results of the optimization are valid
     if not np.isfinite(x).all():
-        raise StatmechFitError('Returned solution vector is nonsensical: x = {0}.'.format(x))
+        raise StatmechFitError("Returned solution vector is nonsensical: x = {0}.".format(x))
     if igo == 8:
-        logging.warning('Maximum number of iterations reached when fitting spectral data for '
-                        '{0}.'.format(molecule.to_smiles()))
+        logging.warning("Maximum number of iterations reached when fitting spectral data for " "{0}.".format(molecule.to_smiles()))
     if igo > 8:
-        logging.warning('A solver error occured when fitting spectral data for {0}.'.format(molecule.to_smiles()))
-    logging.debug('Fitting remaining heat capacity to {0} vibrations and {1} rotations'.format(n_vib, n_rot))
-    logging.debug('The residuals for heat capacity values is {}'.format(fit.evaluate(x)[0]))
+        logging.warning("A solver error occured when fitting spectral data for {0}.".format(molecule.to_smiles()))
+    logging.debug("Fitting remaining heat capacity to {0} vibrations and {1} rotations".format(n_vib, n_rot))
+    logging.debug("The residuals for heat capacity values is {}".format(fit.evaluate(x)[0]))
 
     # Postprocess optimization results
     vib = list(x[0:n_vib])
@@ -252,6 +260,7 @@ def fit_statmech_pseudo_rotors(Tlist, Cvlist, n_vib, n_rot, molecule=None):
 
 
 ################################################################################
+
 
 def fit_statmech_pseudo(Tlist, Cvlist, n_vib, n_rot, molecule=None):
     """
@@ -279,7 +288,7 @@ def fit_statmech_pseudo(Tlist, Cvlist, n_vib, n_rot, molecule=None):
     bounds.append((hr_barr_lower_bound, hr_barr_upper_bound))
 
     # Construct the initial guess
-    x0 = np.zeros(6, np.float64)  # Initial guess
+    x0 = np.zeros(6, float)  # Initial guess
     x0[0] = 300.0
     x0[1] = float(math.floor((n_vib - 1) / 2.0))
     x0[2] = 800.0
@@ -294,21 +303,19 @@ def fit_statmech_pseudo(Tlist, Cvlist, n_vib, n_rot, molecule=None):
 
     # Check that the results of the optimization are valid
     if not np.isfinite(x).all():
-        raise StatmechFitError('Returned solution vector is nonsensical: x = {0}.'.format(x))
+        raise StatmechFitError("Returned solution vector is nonsensical: x = {0}.".format(x))
     if igo == 8:
-        logging.warning('Maximum number of iterations reached when fitting spectral data for '
-                        '{0}.'.format(molecule.to_smiles()))
+        logging.warning("Maximum number of iterations reached when fitting spectral data for " "{0}.".format(molecule.to_smiles()))
     if igo > 8:
-        logging.warning('A solver error occured when fitting spectral data for {0}.'.format(molecule.to_smiles()))
-    logging.debug('Fitting remaining heat capacity to {0} vibrations and {1} rotations'.format(n_vib, n_rot))
-    logging.debug('The residuals for heat capacity values is {}'.format(fit.evaluate(x)[0]))
+        logging.warning("A solver error occured when fitting spectral data for {0}.".format(molecule.to_smiles()))
+    logging.debug("Fitting remaining heat capacity to {0} vibrations and {1} rotations".format(n_vib, n_rot))
+    logging.debug("The residuals for heat capacity values is {}".format(fit.evaluate(x)[0]))
 
     # Postprocess optimization results
     n_vib2 = int(round(x[1]))
     n_vib_3 = n_vib - n_vib2 - 1
     if n_vib2 < 0 or n_vib2 > n_vib - 1 or n_vib_3 < 0 or n_vib_3 > n_vib - 1:
-        raise StatmechFitError('Invalid degeneracies {0} and {1} fitted for '
-                               'pseudo-frequencies.'.format(n_vib2, n_vib_3))
+        raise StatmechFitError("Invalid degeneracies {0} and {1} fitted for " "pseudo-frequencies.".format(n_vib2, n_vib_3))
 
     vib = [x[0]]
     for i in range(n_vib2):
@@ -323,6 +330,7 @@ def fit_statmech_pseudo(Tlist, Cvlist, n_vib, n_rot, molecule=None):
 
 
 ################################################################################
+
 
 def harmonic_oscillator_heat_capacity(T, freq):
     """
@@ -353,10 +361,10 @@ def hindered_rotor_heat_capacity(T, freq, barr):
     in K for the 1D hindered rotor with a frequency `freq` in cm^-1 and a
     barrier height `barr` in cm^-1.
     """
-    x = constants.h * constants.c * 100. * freq / constants.kB / T
+    x = constants.h * constants.c * 100.0 * freq / constants.kB / T
     exp_x = math.exp(x)
     one_minus_exp_x = 1.0 - exp_x
-    z = 0.5 * constants.h * constants.c * 100. * barr / constants.kB / T
+    z = 0.5 * constants.h * constants.c * 100.0 * barr / constants.kB / T
     bb = scipy.special.i1(z) / scipy.special.i0(z)
     return x * x * exp_x / one_minus_exp_x / one_minus_exp_x - 0.5 + z * (z - bb - z * bb * bb)
 
@@ -368,7 +376,7 @@ def hindered_rotor_d_heat_capacity_d_freq(T, freq, barr):
     `Tlist` in K, evaluated at the frequency `freq` in cm^-1 and a barrier
     height `barr` in cm^-1.
     """
-    x = constants.h * constants.c * 100. * freq / constants.kB / T
+    x = constants.h * constants.c * 100.0 * freq / constants.kB / T
     exp_x = math.exp(x)
     one_minus_exp_x = 1.0 - exp_x
     return x * exp_x / one_minus_exp_x / one_minus_exp_x * (2 + x + 2 * x * exp_x / one_minus_exp_x) * x / freq
@@ -381,12 +389,13 @@ def hindered_rotor_d_heat_capacity_d_barr(T, freq, barr):
     `Tlist` in K, evaluated at the frequency `freq` in cm^-1 and a barrier
     height `barr` in cm^-1.
     """
-    z = 0.5 * constants.h * constants.c * 100. * barr / constants.kB / T
+    z = 0.5 * constants.h * constants.c * 100.0 * barr / constants.kB / T
     bb = scipy.special.i1(z) / scipy.special.i0(z)
     return z * (1 - 2 * z * bb + bb * bb + 2 * z * bb * bb * bb) * z / barr
 
 
 ################################################################################
+
 
 class DirectFit(DQED):
     """
@@ -405,10 +414,10 @@ class DirectFit(DQED):
         n_eq = self.Neq
         n_vars = self.Nvars
         n_cons = self.Ncons
-        f = np.zeros((n_eq), np.float64)
-        J = np.zeros((n_eq, n_vars), np.float64)
-        f_cons = np.zeros((n_cons), np.float64)
-        J_cons = np.zeros((n_cons, n_vars), np.float64)
+        f = np.zeros((n_eq), float)
+        J = np.zeros((n_eq, n_vars), float)
+        f_cons = np.zeros((n_cons), float)
+        J_cons = np.zeros((n_cons, n_vars), float)
 
         n_vib = self.n_vib
         n_rot = self.n_rot
@@ -424,10 +433,8 @@ class DirectFit(DQED):
             for n in range(n_vib):
                 J[i, n] = harmonic_oscillator_d_heat_capacity_d_freq(self.Tdata[i], x[n])
             for n in range(n_rot):
-                J[i, n_vib + 2 * n] = hindered_rotor_d_heat_capacity_d_freq(self.Tdata[i], x[n_vib + 2 * n],
-                                                                            x[n_vib + 2 * n + 1])
-                J[i, n_vib + 2 * n + 1] = hindered_rotor_d_heat_capacity_d_barr(self.Tdata[i], x[n_vib + 2 * n],
-                                                                                x[n_vib + 2 * n + 1])
+                J[i, n_vib + 2 * n] = hindered_rotor_d_heat_capacity_d_freq(self.Tdata[i], x[n_vib + 2 * n], x[n_vib + 2 * n + 1])
+                J[i, n_vib + 2 * n + 1] = hindered_rotor_d_heat_capacity_d_barr(self.Tdata[i], x[n_vib + 2 * n], x[n_vib + 2 * n + 1])
 
         return f, J, f_cons, J_cons
 
@@ -451,16 +458,16 @@ class PseudoRotorFit(DQED):
         n_eq = self.Neq
         n_vars = self.Nvars
         n_cons = self.Ncons
-        f = np.zeros((n_eq), np.float64)
-        J = np.zeros((n_eq, n_vars), np.float64)
-        f_cons = np.zeros((n_cons), np.float64)
-        J_cons = np.zeros((n_cons, n_vars), np.float64)
+        f = np.zeros((n_eq), float)
+        J = np.zeros((n_eq, n_vars), float)
+        f_cons = np.zeros((n_cons), float)
+        J_cons = np.zeros((n_cons, n_vars), float)
 
         n_vib = self.n_vib
         n_rot = self.n_rot
 
-        cv = np.zeros((len(self.Tdata), n_vib + 1), np.float64)
-        d_cv = np.zeros((len(self.Tdata), n_vib + 2), np.float64)
+        cv = np.zeros((len(self.Tdata), n_vib + 1), float)
+        d_cv = np.zeros((len(self.Tdata), n_vib + 2), float)
 
         for i in range(len(self.Tdata)):
             for j in range(n_vib):
@@ -504,10 +511,10 @@ class PseudoFit(DQED):
         n_eq = self.Neq
         n_vars = self.Nvars
         n_cons = self.Ncons
-        f = np.zeros((n_eq), np.float64)
-        J = np.zeros((n_eq, n_vars), np.float64)
-        f_cons = np.zeros((n_cons), np.float64)
-        J_cons = np.zeros((n_cons, n_vars), np.float64)
+        f = np.zeros((n_eq), float)
+        J = np.zeros((n_eq, n_vars), float)
+        f_cons = np.zeros((n_cons), float)
+        J_cons = np.zeros((n_cons, n_vars), float)
 
         n_vib = self.n_vib
         n_rot = self.n_rot
